compiles properly, no apparent error but need more investigations with gdb and valgrind, and long runs

+++ include/io.h
@@ -22,11 +22,11 @@ extern FILE *efile;
 extern FILE *stfile;
 
 //pointer to the desired IO function
-void (*write_traj)(ATOM at[], DATA *dat, uint32_t when);
+void (*write_traj)(ATOM at[], DATA *dat, uint64_t when);
 
 void read_xyz(ATOM at[], DATA *dat, FILE *inpf);
-void write_xyz(ATOM at[], DATA *dat, uint32_t when, FILE *outf);
-void write_dcd(ATOM at[], DATA *dat, uint32_t when);
+void write_xyz(ATOM at[], DATA *dat, uint64_t when, FILE *outf);
+void write_dcd(ATOM at[], DATA *dat, uint64_t when);
 
 void write_rst(DATA *dat, SPDAT *spdat, ATOM at[]);
